\name{SmoothImage}
\alias{SmoothImage}
\title{Title}
\description{
    #add description here
}

\usage{
    #main function to show public - main and it's arguments
    smoothimage(a, b)
}

\arguments{
    #write about arguments in detail
    \item{a}{
        The data set you want to whatever
    }
    \item{b}{
        The data set you want to whatever
    }
}

\details{
    #talk about stuff
    Talk about the stuff in general
}

\value{
    #what it will be outputted as
    output form
}

\author{
Norm Matloff <matloff@cs.ucdavis.edu>, you and your info
}

\examples{
    # whatever code you have to show an example
}
