\name{smoothimage}
\alias{smoothimage}
\title{Title}
\description{
    Blah blah bla
}

\usage{
    smoothimage(a, b)
}

\arguments{
    \item{a}{
        The data set you want to whatever
    }
    \item{b}{
        The data set you want to whatever
    }
}

\details{
    Talk about the stauff in general
}

\value{
    output form
}

\author{
Norm Matloff <matloff@cs.ucdavis.edu>, you and your info
}

\examples{
    # whatever code you have to show an example
}
